Create INotificationService
